requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==23.2.0
blinker==1.7.0
chardet==4.0.0
click==8.1.7
Flask==3.0.2
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
multidict==6.0.5
python-dotenv==1.0.1
siegeapi==6.2.5
typing_extensions==4.10.0
Werkzeug==3.0.1
yarl==1.9.4
